feat: expand commandline values from parent dataset
This change uses a feature of ZFSBootMenu's
`org.zfsbootmenu:commandline` ZFS dataset property to expand commandline
values found in parent datasets. In this way it should be easier to set
default kernel parameters that apply to all boot environments, as well
as a way to surgically modify a single boot environment.

With this change, a value of `org.zfsbootmenu:commandline` is set on the
`zroot/ROOT` dataset and the `zroot/ROOT/default` dataset has a value of
`org.zfsbootmenu:commandline=%{parent}` which expands all values found
in the `org.zfsbootmenu:commandline` from the `zroot/ROOT` dataset. At
install time `rw` is the value at `zroot/ROOT`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>