fix: add temp user when delegating user ZFS dataset perms
This change fixes an issue where the `zfs allow` is happening outside
the chroot and therefore is checking for a *local* user on from the live
ISO environment. As per the [`zfs allow`] documentation, a UID cannot be
provided and the user must be referred to by their username. Therefore,
this change temporarily adds a local user with the same basic
information before running the `zfs allow` and then promptly removes
that user.

[`zfs allow`]: https://openzfs.github.io/openzfs-docs/man/v2.2/8/zfs-allow.8.html

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>